Take out flex timetable into a reusable module
