+++ app/scripts/venues/views/VenueInformationView.js
@@ -6,11 +6,15 @@ var Marionette = require('backbone.marionette');
 var template = require('../templates/venue_information.hbs');
 var Backbone = require('backbone');
 var timify = require('../../common/utils/timify');
+var TimetableFlexView = require('../../timetable_flex/views/TimetableFlexView');
 
 module.exports = Marionette.LayoutView.extend({
   template: template,
   initialize: function () {
     
+  },
+  regions: {
+    venueTimetableRegion: '.nm-venue-information-timetable'
   },
   events: {
     'keypress input[type=text]': 'processKey',
@@ -30,47 +34,24 @@ module.exports = Marionette.LayoutView.extend({
     this.showAvailabilityForVenue(searchText);
   },
   showAvailabilityForVenue: function (venue) {
+    if (venue !== '') {
+      Backbone.history.navigate('venues/' + venue);
+    }
+
     var selectedVenue = null;
     if (this.model.get('venuesList').indexOf(venue) > -1) {
       selectedVenue = this.model.get('venues')[venue];
     }
+
     this.model.set('selectedVenue', selectedVenue);
-    this.model.set('selectedVenueName', venue);
-    _.each(selectedVenue, function (day) {
-      var range = _.map(_.range(timify.convertTimeToIndex('0800'), 
-                                timify.convertTimeToIndex('2400')), function () {
-        return {
-          width: 1,
-          module: ''
-        }
-      });
-
-      _.each(day.classes, function (cls) {
-        var eightAmIndex = timify.convertTimeToIndex('0800');
-        var startIndex = timify.convertTimeToIndex(cls.StartTime) - eightAmIndex;
-        var endIndex = timify.convertTimeToIndex(cls.EndTime) - eightAmIndex;
-        var width = endIndex - startIndex;
-        range[startIndex] = {
-          width: width,
-          module: cls.ModuleCode,
-          class: 'nm-flex-occupied'
-        };
-      });
-
-      var finalRange = [];
-      var step;
-      for (var i = 0; i < range.length; i += step) {
-        step = range[i].width;
-        finalRange.push(range[i]);
-      }
-      day.timetable = finalRange;
-      console.log(_.reduce(_.pluck(finalRange, 'width'), function (memo, num) {
-        return memo + num;
-      }));
-    });
-    if (venue !== '') {
-      Backbone.history.navigate('venue/' + venue);
-    }
     this.render();
+
+    var venueTimetableModel = new Backbone.Model({
+      venueDayAvailability: selectedVenue
+    });
+
+    this.venueTimetableRegion.show(new TimetableFlexView({
+      model: venueTimetableModel
+    }));
   }
 });
